requirements.txt
tensorflow
numpy
pandas
matplotlib
seaborn
scikit-learn
opencv-python
pillow
streamlit
tqdm